Refactored the countdown timer code so it can be modular and be used in newitems.jsx and exploreitems.jsx as a Countdown component
